--- src/test/java/cart/dto/ProductModificationRequestTest.java
@@ -1,68 +1,0 @@
-package cart.dto;
-
-import cart.product.dto.ProductModificationRequest;
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.NullAndEmptySource;
-import org.junit.jupiter.params.provider.NullSource;
-
-import javax.validation.ConstraintViolation;
-import java.util.Set;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-@SuppressWarnings("NonAsciiCharacters")
-class ProductModificationRequestTest extends ValidationTest {
-
-    @Test
-    void 제대로_된_데이터_형식이_들어오면_성공한다() {
-        final ProductModificationRequest request = new ProductModificationRequest(dummyId, dummyName, dummyImageUrl, dummyPrice);
-
-        final Set<ConstraintViolation<ProductModificationRequest>> constraintViolations = validator.validate(request);
-
-        assertThat(constraintViolations.isEmpty()).isTrue();
-    }
-
-    @NullSource
-    @ParameterizedTest
-    void 아이디기_없으면_에러를_반환한다(final Long input) {
-        final ProductModificationRequest request = new ProductModificationRequest(input, dummyName, dummyImageUrl, dummyPrice);
-
-        final Set<ConstraintViolation<ProductModificationRequest>> constraintViolations = validator.validate(request);
-
-        assertThat(constraintViolations.isEmpty()).isFalse();
-    }
-
-    @NullAndEmptySource
-    @ParameterizedTest
-    void 이름이_없으면_에러를_반환한다(final String input) {
-        final ProductModificationRequest request = new ProductModificationRequest(dummyId, input, dummyImageUrl, dummyPrice);
-
-        final Set<ConstraintViolation<ProductModificationRequest>> constraintViolations = validator.validate(request);
-
-        assertThat(constraintViolations.isEmpty()).isFalse();
-    }
-
-    @NullAndEmptySource
-    @ParameterizedTest
-    void 이미지_경로가_없으면_에러를_반환한다(final String input) {
-        final ProductModificationRequest request = new ProductModificationRequest(dummyId, dummyName, input, dummyPrice);
-
-        final Set<ConstraintViolation<ProductModificationRequest>> constraintViolations = validator.validate(request);
-
-        assertThat(constraintViolations.isEmpty()).isFalse();
-    }
-
-    @NullSource
-    @ParameterizedTest
-    void 가격이_없으면_에러를_반환한다(final Integer input) {
-        final ProductModificationRequest request = new ProductModificationRequest(dummyId, dummyName, dummyImageUrl, input);
-
-        final Set<ConstraintViolation<ProductModificationRequest>> constraintViolations = validator.validate(request);
-
-        assertThat(constraintViolations.isEmpty()).isFalse();
-    }
-}
